# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MMJ Payroll System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /hotels:
    get:
      operationId: listHotels
      tags: [hotels]
      summary: List all hotels/worksites
      parameters:
        - name: region
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of hotels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hotel"
    post:
      operationId: createHotel
      tags: [hotels]
      summary: Create a hotel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHotelBody"
      responses:
        "201":
          description: Created hotel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hotel"
  /hotels/{id}:
    get:
      operationId: getHotel
      tags: [hotels]
      summary: Get a hotel by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Hotel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hotel"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateHotel
      tags: [hotels]
      summary: Update a hotel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHotelBody"
      responses:
        "200":
          description: Updated hotel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hotel"
    delete:
      operationId: deleteHotel
      tags: [hotels]
      summary: Delete a hotel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Pay Periods ─────────────────────────────────────────────────────────────
...
  /import/analyze:
    post:
      operationId: analyzeImport
      tags: [payPeriods]
      summary: Analyze uploaded Excel workbook and return sheet summary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeImportBody"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportAnalysis"
  /import/confirm:
    post:
      operationId: confirmImport
      tags: [payPeriods]
      summary: Confirm and execute import into a pay period
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmImportBody"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"

  # ─── Business Profiles ───────────────────────────────────────────────────────
  /business-profiles:
    get:
      operationId: listBusinessProfiles
      tags: [businessProfiles]
      summary: List all business profiles
      responses:
        "200":
          description: List of profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BusinessProfile"
    post:
      operationId: createBusinessProfile
      tags: [businessProfiles]
      summary: Create a business profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBusinessProfileBody"
      responses:
        "201":
          description: Created profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfile"
  /business-profiles/{id}:
    get:
      operationId: getBusinessProfile
      tags: [businessProfiles]
      summary: Get a business profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfile"
    patch:
      operationId: updateBusinessProfile
      tags: [businessProfiles]
      summary: Update a business profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBusinessProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfile"
    delete:
      operationId: deleteBusinessProfile
      tags: [businessProfiles]
      summary: Delete a business profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /business-profiles/{id}/set-default:
    post:
      operationId: setDefaultBusinessProfile
      tags: [businessProfiles]
      summary: Set a profile as the default
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfile"

  # ─── Logos ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Hotel:
      type: object
      properties:
        id:
          type: integer
        renderDbId:
          type: ["string", "null"]
        name:
          type: string
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
        region:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        isActive:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, isActive, createdAt, updatedAt]
    CreateHotelBody:
      type: object
      properties:
        name:
          type: string
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
        region:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required: [name]
    UpdateHotelBody:
      type: object
      properties:
        name:
          type: string
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
        region:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        isActive:
          type: boolean
        notes:
          type: ["string", "null"]
...
    AnalyzeImportBody:
      type: object
      properties:
        fileData:
          type: string
          description: Base64-encoded Excel file
        fileName:
          type: string
      required: [fileData, fileName]
    ImportAnalysis:
      type: object
      properties:
        fileName:
          type: string
        sheets:
          type: array
          items:
            $ref: "#/components/schemas/SheetSummary"
        detectedPeriodName:
          type: ["string", "null"]
        detectedStartDate:
          type: ["string", "null"]
        detectedEndDate:
          type: ["string", "null"]
        workerCount:
          type: integer
        entryCount:
          type: integer
      required: [fileName, sheets, workerCount, entryCount]
    SheetSummary:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        rowCount:
          type: integer
        recognized:
          type: boolean
      required: [name, type, rowCount, recognized]
    ConfirmImportBody:
      type: object
      properties:
        fileData:
          type: string
          description: Base64-encoded Excel file
        periodId:
          type: ["integer", "null"]
          description: Existing period to import into (null = create new)
        periodName:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        sheetsToImport:
          type: array
          items:
            type: string
        replaceExisting:
          type: boolean
      required: [fileData, sheetsToImport]
    ImportResult:
      type: object
      properties:
        periodId:
          type: integer
        periodName:
          type: string
        entriesImported:
          type: integer
        workersCreated:
          type: integer
        warnings:
          type: array
          items:
            type: string
      required: [periodId, periodName, entriesImported, workersCreated, warnings]
    BusinessProfile:
      type: object
      properties:
        id:
          type: integer
        businessName:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        hstNumber:
          type: ["string", "null"]
        logoId:
          type: ["integer", "null"]
        logoUrl:
          type: ["string", "null"]
        isDefault:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, businessName, isDefault, createdAt, updatedAt]
    CreateBusinessProfileBody:
      type: object
      properties:
        businessName:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        hstNumber:
          type: ["string", "null"]
        logoId:
          type: ["integer", "null"]
        isDefault:
          type: boolean
        notes:
          type: ["string", "null"]
      required: [businessName]
    UpdateBusinessProfileBody:
      type: object
      properties:
        businessName:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        hstNumber:
          type: ["string", "null"]
        logoId:
          type: ["integer", "null"]
        isDefault:
          type: boolean
        notes:
          type: ["string", "null"]
...