# lib/api-zod/src/generated/types/updateHotelBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MMJ Payroll System API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateHotelBody {
  name?: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  province?: string | null;
  /** @nullable */
  region?: string | null;
  /** @nullable */
  contactName?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  isActive?: boolean;
  /** @nullable */
  notes?: string | null;
}